+++ Sources/ApptentiveKit/Interactions/MessageCenter/Default UI/MessageComposeView.swift
@@ -11,15 +11,23 @@ import UIKit
 /// Implements `intrinsicContentSize` without interfering with the autolayout of the compose view.
 class MessageCenterComposeContainerView: UIView {
     let composeView: MessageCenterComposeView
+    let separatorView: UIView
+    var separatorHeightConstraint = NSLayoutConstraint()
 
-    override init(frame: CGRect) {
-        self.composeView = MessageCenterComposeView(frame: frame)
+    init(composeView: MessageCenterComposeView) {
+        self.composeView = composeView
+        self.separatorView = UIView(frame: .zero)
 
-        super.init(frame: frame)
+        super.init(frame: .zero)
 
         self.addSubview(self.composeView)
+        self.addSubview(self.separatorView)
+
         self.translatesAutoresizingMaskIntoConstraints = false
         self.backgroundColor = .apptentiveMessageCenterComposeBoxBackground
+
+        self.separatorView.translatesAutoresizingMaskIntoConstraints = false
+        self.separatorView.backgroundColor = .apptentiveMessageCenterComposeBoxSeparator
 
         self.setUpConstraints()
     }
@@ -39,24 +47,36 @@ class MessageCenterComposeContainerView: UIView {
     private func setUpConstraints() {
         self.composeView.translatesAutoresizingMaskIntoConstraints = false
 
+        self.separatorHeightConstraint = self.separatorView.heightAnchor.constraint(equalToConstant: 1.0)
+
         NSLayoutConstraint.activate([
-            self.topAnchor.constraint(equalTo: self.composeView.topAnchor),
+            self.separatorView.topAnchor.constraint(equalTo: self.topAnchor),
+            self.separatorView.leadingAnchor.constraint(equalTo: self.leadingAnchor),
+            self.separatorView.trailingAnchor.constraint(equalTo: self.trailingAnchor),
+            self.separatorHeightConstraint,
+
+            self.composeView.topAnchor.constraint(equalToSystemSpacingBelow: self.topAnchor, multiplier: 1),
             self.leadingAnchor.constraint(equalTo: self.composeView.leadingAnchor),
             self.trailingAnchor.constraint(equalTo: self.composeView.trailingAnchor),
             self.safeAreaLayoutGuide.bottomAnchor.constraint(equalTo: self.composeView.bottomAnchor),
         ])
+    }
+
+    override func layoutSubviews() {
+        super.layoutSubviews()
+
+        let fineLineWidth = 1.0 / self.traitCollection.displayScale
+        self.separatorHeightConstraint.constant = fineLineWidth
     }
 }
 
 class MessageCenterComposeView: UIView {
     let textView: UITextView
     let placeholderLabel: UILabel
-    let separatorView: UIView
     let sendButton: UIButton
     var placeholderWidthConstraint: NSLayoutConstraint?
     var textViewHeightConstraint: NSLayoutConstraint?
     var textViewHeightLimitConstraint: NSLayoutConstraint?
-    var separatorHeightConstraint: NSLayoutConstraint?
     let attachmentButton: UIButton
     let attachmentStackView: UIStackView
 
@@ -64,7 +84,6 @@ class MessageCenterComposeView: UIView {
         self.textView = UITextView(frame: frame)
         self.placeholderLabel = UILabel(frame: frame)
         self.sendButton = UIButton(frame: frame)
-        self.separatorView = UIView(frame: frame)
         self.attachmentButton = UIButton(frame: frame)
         self.attachmentStackView = UIStackView(frame: frame)
 
@@ -72,7 +91,6 @@ class MessageCenterComposeView: UIView {
 
         self.addSubview(self.textView)
         self.addSubview(self.sendButton)
-        self.addSubview(self.separatorView)
         self.addSubview(self.attachmentButton)
         self.addSubview(self.attachmentStackView)
 
@@ -105,9 +123,6 @@ class MessageCenterComposeView: UIView {
         self.textView.returnKeyType = .default
         self.textView.accessibilityIdentifier = "messageTextView"
 
-        self.separatorView.translatesAutoresizingMaskIntoConstraints = false
-        self.separatorView.backgroundColor = .apptentiveMessageCenterComposeBoxSeparator
-
         self.textView.addSubview(self.placeholderLabel)
         self.placeholderLabel.isAccessibilityElement = false
         self.placeholderLabel.translatesAutoresizingMaskIntoConstraints = false
@@ -128,6 +143,7 @@ class MessageCenterComposeView: UIView {
         }
 
         self.textView.accessibilityIdentifier = "composeTextView"
+        self.accessibilityElements = [self.attachmentButton, self.textView, self.sendButton, self.attachmentStackView]
 
         self.updatePlaceholderConstraints()
     }
@@ -162,6 +178,7 @@ class MessageCenterComposeView: UIView {
             self.sendButton.setPreferredSymbolConfiguration(.init(pointSize: 32.0), forImageIn: .normal)
         }
         self.sendButton.setImage(.apptentiveMessageSendButton, for: .normal)
+        self.sendButton.contentEdgeInsets = UIEdgeInsets(top: 6, left: 6, bottom: 6, right: 6)
         self.sendButton.accessibilityIdentifier = "sendButton"
     }
 
@@ -171,6 +188,7 @@ class MessageCenterComposeView: UIView {
             self.attachmentButton.setPreferredSymbolConfiguration(.init(pointSize: 32.0), forImageIn: .normal)
         }
         self.attachmentButton.setImage(.apptentiveMessageAttachmentButton, for: .normal)
+        self.attachmentButton.contentEdgeInsets = UIEdgeInsets(top: 6, left: 6, bottom: 6, right: 6)
         self.attachmentButton.accessibilityIdentifier = "attachmentButton"
     }
 
@@ -181,24 +199,17 @@ class MessageCenterComposeView: UIView {
     }
 
     private func setUpConstraints() {
-        self.separatorHeightConstraint = self.separatorView.heightAnchor.constraint(equalToConstant: 1.0)
-
         self.textViewHeightConstraint = self.textView.heightAnchor.constraint(greaterThanOrEqualToConstant: 18.0)
-        self.textViewHeightConstraint?.priority = .defaultHigh
+        self.textViewHeightConstraint?.priority = .init(751)
         self.textViewHeightLimitConstraint = self.textView.heightAnchor.constraint(lessThanOrEqualToConstant: 100)
 
         NSLayoutConstraint.activate(
             [
-                self.separatorView.topAnchor.constraint(equalTo: self.topAnchor),
-                self.separatorView.leadingAnchor.constraint(equalTo: self.leadingAnchor),
-                self.separatorView.trailingAnchor.constraint(equalTo: self.trailingAnchor),
-                self.separatorHeightConstraint,
-
                 self.attachmentButton.centerYAnchor.constraint(equalTo: self.textView.centerYAnchor),
                 self.attachmentButton.leadingAnchor.constraint(equalToSystemSpacingAfter: self.safeAreaLayoutGuide.leadingAnchor, multiplier: 1),
                 self.attachmentButton.heightAnchor.constraint(equalTo: self.attachmentButton.widthAnchor),
 
-                self.textView.topAnchor.constraint(equalToSystemSpacingBelow: self.topAnchor, multiplier: 1.0),
+                self.textView.topAnchor.constraint(equalTo: self.topAnchor),
                 self.textView.leadingAnchor.constraint(equalToSystemSpacingAfter: self.attachmentButton.trailingAnchor, multiplier: 1.0),
                 self.sendButton.leadingAnchor.constraint(equalToSystemSpacingAfter: self.textView.trailingAnchor, multiplier: 1.0),
 
@@ -220,6 +231,5 @@ class MessageCenterComposeView: UIView {
 
         let fineLineWidth = 1.0 / self.traitCollection.displayScale
         self.textView.layer.borderWidth = fineLineWidth
-        self.separatorHeightConstraint?.constant = fineLineWidth
     }
 }