+++ Sources/ApptentiveKit/Interactions/MessageCenter/Default UI/MessageReceivedCell.swift
@@ -9,6 +9,7 @@
 import UIKit
 
 class MessageReceivedCell: UITableViewCell {
+    let stackView: UIStackView
     let messageText: UITextView
     let profileImageView: ApptentiveImageView
     let senderLabel: UILabel
@@ -17,6 +18,7 @@ class MessageReceivedCell: UITableViewCell {
     let attachmentStackView: UIStackView
 
     override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
+        self.stackView = UIStackView(frame: .zero)
         self.messageText = UITextView(frame: .zero)
         self.dateLabel = UILabel(frame: .zero)
         self.senderLabel = UILabel(frame: .zero)
@@ -27,11 +29,13 @@ class MessageReceivedCell: UITableViewCell {
         super.init(style: .default, reuseIdentifier: reuseIdentifier)
 
         self.contentView.addSubview(self.bubbleImageView)
-        self.contentView.addSubview(self.senderLabel)
-        self.contentView.addSubview(self.messageText)
-        self.contentView.addSubview(self.dateLabel)
+        self.contentView.addSubview(self.stackView)
         self.contentView.addSubview(self.profileImageView)
-        self.contentView.addSubview(self.attachmentStackView)
+
+        self.stackView.addArrangedSubview(self.senderLabel)
+        self.stackView.addArrangedSubview(self.messageText)
+        self.stackView.addArrangedSubview(self.attachmentStackView)
+        self.stackView.addArrangedSubview(self.dateLabel)
 
         setupViews()
     }
@@ -43,12 +47,12 @@ class MessageReceivedCell: UITableViewCell {
     private func setupViews() {
         self.backgroundColor = .clear
 
-        self.messageText.textColor = .apptentiveMessageLabelInbound
-        self.senderLabel.textColor = .apptentiveMessageLabelInbound
-        self.dateLabel.textColor = .apptentiveMessageLabelInbound
+        self.stackView.translatesAutoresizingMaskIntoConstraints = false
+        self.stackView.axis = .vertical
+        self.stackView.alignment = .leading
+        self.stackView.spacing = 8
 
         self.bubbleImageView.translatesAutoresizingMaskIntoConstraints = false
-        self.bubbleImageView.setContentHuggingPriority(.defaultLow, for: .vertical)
         self.bubbleImageView.tintColor = .apptentiveMessageBubbleInbound
 
         self.profileImageView.translatesAutoresizingMaskIntoConstraints = false
@@ -56,22 +60,26 @@ class MessageReceivedCell: UITableViewCell {
 
         self.dateLabel.translatesAutoresizingMaskIntoConstraints = false
         self.dateLabel.numberOfLines = 0
+        self.dateLabel.textColor = .apptentiveMessageLabelInbound
         self.dateLabel.font = .apptentiveMessageDateLabel
         self.dateLabel.textAlignment = .left
         self.dateLabel.adjustsFontForContentSizeCategory = true
 
-        self.messageText.font = .apptentiveMessageLabel
         self.messageText.translatesAutoresizingMaskIntoConstraints = false
         self.messageText.isEditable = false
         self.messageText.isScrollEnabled = false
         self.messageText.backgroundColor = .apptentiveMessageBubbleInbound
-        self.messageText.sizeToFit()
-        self.messageText.setContentHuggingPriority(.defaultHigh, for: .vertical)
+        self.messageText.textColor = .apptentiveMessageLabelInbound
+        self.messageText.font = .apptentiveMessageLabel
+        self.messageText.textContainerInset = UIEdgeInsets(top: -3, left: 0, bottom: 0, right: 0)
+        self.messageText.textContainer.lineFragmentPadding = 0
         self.messageText.adjustsFontForContentSizeCategory = true
         self.messageText.dataDetectorTypes = UIDataDetectorTypes.all
+        self.messageText.isAccessibilityElement = true  // Make navigable via full keyboard access.
 
         self.senderLabel.translatesAutoresizingMaskIntoConstraints = false
         self.senderLabel.numberOfLines = 0
+        self.senderLabel.textColor = .apptentiveMessageLabelInbound
         self.senderLabel.font = .apptentiveSenderLabel
         self.senderLabel.textAlignment = .left
         self.senderLabel.adjustsFontForContentSizeCategory = true
@@ -84,29 +92,29 @@ class MessageReceivedCell: UITableViewCell {
     }
 
     private func setConstraints() {
+        NSLayoutConstraint(
+            item: bubbleImageView,
+            attribute: .trailing,
+            relatedBy: .equal,
+            toItem: self.contentView,
+            attribute: .trailing,
+            multiplier: 0.75,
+            constant: 0
+        ).isActive = true
+
+        let bottomConstraint = self.bubbleImageView.bottomAnchor.constraint(equalTo: self.stackView.bottomAnchor, constant: 15)
+        bottomConstraint.priority = .init(751)
+
         NSLayoutConstraint.activate(
             [
                 self.bubbleImageView.topAnchor.constraint(equalTo: self.contentView.topAnchor, constant: 10),
                 self.bubbleImageView.leadingAnchor.constraint(equalTo: self.contentView.leadingAnchor, constant: 40),
                 self.contentView.bottomAnchor.constraint(equalTo: self.bubbleImageView.bottomAnchor, constant: 10),
-                self.contentView.trailingAnchor.constraint(greaterThanOrEqualTo: self.bubbleImageView.trailingAnchor, constant: 30),
 
-                self.senderLabel.topAnchor.constraint(equalTo: self.bubbleImageView.topAnchor, constant: 10),
-                self.senderLabel.leadingAnchor.constraint(equalTo: self.bubbleImageView.leadingAnchor, constant: 20),
-                self.bubbleImageView.trailingAnchor.constraint(equalTo: self.senderLabel.trailingAnchor, constant: 10),
-
-                self.messageText.topAnchor.constraint(equalTo: self.senderLabel.bottomAnchor, constant: 10),
-                self.messageText.leadingAnchor.constraint(equalTo: self.senderLabel.leadingAnchor),
-                self.messageText.trailingAnchor.constraint(equalTo: self.senderLabel.trailingAnchor),
-
-                self.attachmentStackView.topAnchor.constraint(equalTo: self.messageText.bottomAnchor, constant: 5),
-                self.attachmentStackView.leadingAnchor.constraint(equalTo: self.messageText.leadingAnchor),
-                self.attachmentStackView.trailingAnchor.constraint(lessThanOrEqualTo: self.messageText.trailingAnchor),
-
-                self.dateLabel.topAnchor.constraint(equalTo: self.attachmentStackView.bottomAnchor, constant: 5),
-                self.dateLabel.leadingAnchor.constraint(equalTo: self.senderLabel.leadingAnchor),
-                self.dateLabel.trailingAnchor.constraint(equalTo: self.senderLabel.trailingAnchor),
-                self.bubbleImageView.bottomAnchor.constraint(equalToSystemSpacingBelow: self.dateLabel.bottomAnchor, multiplier: 2.0),
+                self.stackView.topAnchor.constraint(equalTo: self.bubbleImageView.topAnchor, constant: 15),
+                self.stackView.leadingAnchor.constraint(equalTo: self.bubbleImageView.leadingAnchor, constant: 27),
+                self.bubbleImageView.trailingAnchor.constraint(equalTo: self.stackView.trailingAnchor, constant: 15),
+                bottomConstraint,
 
                 self.profileImageView.leadingAnchor.constraint(equalTo: self.contentView.leadingAnchor, constant: 10),
                 self.profileImageView.bottomAnchor.constraint(equalTo: self.bubbleImageView.bottomAnchor),