+++ Sources/ApptentiveKit/Model/AttachmentManager.swift
@@ -62,7 +62,11 @@ class AttachmentManager: AttachmentURLProviding {
 
     func removeStorage(for attachment: MessageList.Message.Attachment) throws {
         if let attachmentURL = self.url(for: attachment) {
-            try self.fileManager.removeItem(at: attachmentURL)
+            if self.fileManager.fileExists(atPath: attachmentURL.path) {
+                try self.fileManager.removeItem(at: attachmentURL)
+            } else {
+                ApptentiveLogger.default.error("File does not exist at attachment URL path when attempting to remove from storage.")
+            }
         }  // else it likely doesn't have a sidecar file.
     }
 
@@ -198,7 +202,10 @@ class AttachmentManager: AttachmentURLProviding {
         let cancellable = self.requestor.download(remoteURL) { tempURL, response, error in
             completion(
                 Result(catching: {
+                    self.progressObservations.removeValue(forKey: remoteURL)?.invalidate()
+
                     if let error = error {
+                        progress?(0)
                         throw error
                     }
 
@@ -208,7 +215,6 @@ class AttachmentManager: AttachmentURLProviding {
                     ApptentiveLogger.attachments.debug("Trying to write attachment data from \(remoteURL.path) to \(localURL.path).")
                     try self.fileManager.moveItem(at: tempURL, to: localURL)
                     ApptentiveLogger.attachments.debug("Successfully wrote attachment data to \(localURL.path).")
-                    self.progressObservations.removeValue(forKey: remoteURL)?.invalidate()
 
                     return localURL
                 })